Introduce symbol versioning
This adds the option to compile with symbol versioning.  The symbol list
is obtained from the header files by filtering those marked with the
LIBSSH_API modifier.  Such symbols are used as input to generate the
version script used by the linker.  The version script is automatically
updated as new symbols marked with LIBSSH_API are added to the header
files.  If any symbol is removed, the build will fail due to break in
the ABI.

Symbol versioning is enabled by default, but disabled in non UNIX like
platforms. It can be disabled by passing "-DWITH_SYMBOL_VERSIONING=OFF"
option to cmake or "-withoutsymbolversioning" to "obj/build_make.sh".