Run scripts in build time
Parts of the previous scripts should run in build time.  The relevant
parts were moved to separated scripts and run in build time.  Added a
custom command to copy the generated map file back to the source
directory.  Added option to mark a release as final, preventing it to be
modified later.