Use functions instead of macros for mutexes
The macros used for mutex were replaced by functions defined in
threads.h. The actual used implementation is defined in configuration
time. In the new implementation, an initializer is called by the
constructor when the library is loaded instead if using the static
initializer. This also checks the return value of mutex functions,
exiting in case of failure.